refactor CLI stuff into some classes for easier testing
